requirements.txt
Django==2.0.3
django-cors-headers==2.0.0
django-debug-toolbar==1.7
django-rest-swagger==2.1.2
djangorestframework==3.7.7
drf-yasg==1.7.1
drfdocs==0.0.11
drf-autodocs==0.4.4
flex==6.13.1
ipython==5.3.0
requests==2.13.0
django-extensions==2.0.6
swagger-spec-validator==2.1.0

pytest-cov==2.5.1
pytest==3.5.0
pytest-django==3.1.2
pytest-mock==1.7.1
